format

+++ apps/foundry_etherscan-verification/src/Counter.sol
@@ -6,27 +6,27 @@ import {ExternalMath} from "./ExternalMath.sol";
 import {Oracle} from "./Oracle.sol";
 
 contract Counter {
-    using InternalMath for uint;
-    using ExternalMath for uint;
+    using InternalMath for uint256;
+    using ExternalMath for uint256;
 
-    uint public x;
+    uint256 public x;
     Oracle public oracle;
 
-    event Increment(uint by);
-    event Multiply(uint by);
+    event Increment(uint256 by);
+    event Multiply(uint256 by);
 
     constructor(Oracle _oracle) {
         oracle = _oracle;
     }
 
     function inc() public {
-        uint by = oracle.by();
+        uint256 by = oracle.by();
         x = x.add(by);
         emit Increment(by);
     }
 
     function mul() public {
-        uint by = oracle.by();
+        uint256 by = oracle.by();
         x = x.mul(by);
         emit Multiply(by);
     }
