+++ apps/foundry_etherscan-verification/src/ExternalMath.sol
@@ -2,7 +2,7 @@
 pragma solidity ^0.8.28;
 
 library ExternalMath {
-    function mul(uint a, uint b) external pure returns (uint) {
+    function mul(uint256 a, uint256 b) external pure returns (uint256) {
         return a * b;
     }
 }
