+++ apps/foundry_etherscan-verification/src/Oracle.sol
@@ -2,11 +2,11 @@
 pragma solidity ^0.8.28;
 
 contract Oracle {
-    uint public by;
+    uint256 public by;
 
-    event Updated(uint by);
+    event Updated(uint256 by);
 
-    function set(uint _by) public {
+    function set(uint256 _by) public {
         by = _by;
         emit Updated(_by);
     }
